+++ lib/Steps/ChooseIntegration.ts
@@ -1,6 +1,5 @@
 import type { Answers } from 'inquirer';
 import { prompt } from 'inquirer';
-import * as _ from 'lodash';
 import { dim } from 'picocolors';
 
 import {
@@ -15,7 +14,9 @@ import { Electron } from './Integrations/Electron';
 import { NextJsShim } from './Integrations/NextJsShim';
 import { ReactNative } from './Integrations/ReactNative';
 import { SourceMapsShim } from './Integrations/SourceMapsShim';
+import { Apple } from './Integrations/Apple';
 import { SvelteKitShim } from './Integrations/SvelteKitShim';
+import { hasPackageInstalled } from '../../src/utils/package-json';
 import { Android } from './Integrations/Android';
 
 let projectPackage: any = {};
@@ -55,6 +56,9 @@ export class ChooseIntegration extends BaseStep {
       case Integration.sourcemaps:
         integration = new SourceMapsShim(this._argv);
         break;
+      case Integration.ios:
+        integration = new Apple(this._argv);
+        break;
       case Integration.reactNative:
       default:
         integration = new ReactNative(sanitizeUrl(this._argv));
@@ -65,12 +69,22 @@ export class ChooseIntegration extends BaseStep {
   }
 
   public tryDetectingIntegration(): Integration | undefined {
-    if (_.has(projectPackage, 'dependencies.react-native')) {
+    if (hasPackageInstalled('react-native', projectPackage)) {
       return Integration.reactNative;
     }
-    if (_.has(projectPackage, 'dependencies.cordova')) {
+    if (hasPackageInstalled('cordova', projectPackage)) {
       return Integration.cordova;
     }
+    if (hasPackageInstalled('electron', projectPackage)) {
+      return Integration.electron;
+    }
+    if (hasPackageInstalled('next', projectPackage)) {
+      return Integration.nextjs;
+    }
+    if (hasPackageInstalled('@sveltejs/kit', projectPackage)) {
+      return Integration.sveltekit;
+    }
+
     return;
   }