+++ src/sourcemaps/sourcemaps-wizard.ts
@@ -5,16 +5,17 @@ import * as Sentry from '@sentry/node';
 
 import {
   abortIfCancelled,
-  askForProjectSelection,
-  askForSelfHosted,
-  askForWizardLogin,
   confirmContinueEvenThoughNoGitRepo,
+  detectPackageManager,
+  SENTRY_DOT_ENV_FILE,
   printWelcome,
+  SENTRY_CLI_RC_FILE,
+  getOrAskForProjectData,
 } from '../utils/clack-utils';
 import { isUnicodeSupported } from '../utils/vendor/is-unicorn-supported';
 import { SourceMapUploadToolConfigurationOptions } from './tools/types';
 import { configureVitePlugin } from './tools/vite';
-import { configureSentryCLI } from './tools/sentry-cli';
+import { setupNpmScriptInCI, configureSentryCLI } from './tools/sentry-cli';
 import { configureWebPackPlugin } from './tools/webpack';
 import { configureTscSourcemapGenerationFlow } from './tools/tsc';
 import { configureRollupPlugin } from './tools/rollup';
@@ -22,71 +23,116 @@ import { configureEsbuildPlugin } from './tools/esbuild';
 import { WizardOptions } from '../utils/types';
 import { configureCRASourcemapGenerationFlow } from './tools/create-react-app';
 import { ensureMinimumSdkVersionIsInstalled } from './utils/sdk-version';
-import { traceStep } from '../telemetry';
-
-type SupportedTools =
-  | 'webpack'
-  | 'vite'
-  | 'rollup'
-  | 'esbuild'
-  | 'tsc'
-  | 'sentry-cli'
-  | 'create-react-app';
+import { traceStep, withTelemetry } from '../telemetry';
+import { URL } from 'url';
+import { checkIfMoreSuitableWizardExistsAndAskForRedirect } from './utils/other-wizards';
+import { configureAngularSourcemapGenerationFlow } from './tools/angular';
+import { detectUsedTool, SupportedTools } from './utils/detect-tool';
+import { configureNextJsSourceMapsUpload } from './tools/nextjs';
 
 export async function runSourcemapsWizard(
   options: WizardOptions,
 ): Promise<void> {
+  return withTelemetry(
+    {
+      enabled: options.telemetryEnabled,
+      integration: 'sourcemaps',
+    },
+    () => runSourcemapsWizardWithTelemetry(options),
+  );
+}
+
+async function runSourcemapsWizardWithTelemetry(
+  options: WizardOptions,
+): Promise<void> {
   printWelcome({
     wizardName: 'Sentry Source Maps Upload Configuration Wizard',
-    message:
-      'This wizard will help you upload source maps to Sentry as part of your build.\nThank you for using Sentry :)\n\n(This setup wizard sends telemetry data and crash reports to Sentry.\nYou can turn this off by running the wizard with the `--disable-telemetry` flag.)',
+    message: `This wizard will help you upload source maps to Sentry as part of your build.
+Thank you for using Sentry :)${
+      options.telemetryEnabled
+        ? `
+
+(This setup wizard sends telemetry data and crash reports to Sentry.
+You can turn this off by running the wizard with the '--disable-telemetry' flag.)`
+        : ''
+    }`,
     promoCode: options.promoCode,
   });
+
+  const moreSuitableWizard = await traceStep(
+    'check-framework-wizard',
+    checkIfMoreSuitableWizardExistsAndAskForRedirect,
+  );
+  if (moreSuitableWizard) {
+    await traceStep('run-framework-wizard', () => moreSuitableWizard(options));
+    return;
+  }
 
   await traceStep('detect-git', confirmContinueEvenThoughNoGitRepo);
 
   await traceStep('check-sdk-version', ensureMinimumSdkVersionIsInstalled);
 
-  const { url: sentryUrl, selfHosted } = await traceStep(
-    'ask-self-hosted',
-    () => askForSelfHosted(options.url),
-  );
+  const { selfHosted, selectedProject, sentryUrl, authToken } =
+    await getOrAskForProjectData(options);
 
-  const { projects, apiKeys } = await traceStep('login', () =>
-    askForWizardLogin({
-      promoCode: options.promoCode,
-      url: sentryUrl,
-    }),
-  );
-
-  const selectedProject = await traceStep('select-project', () =>
-    askForProjectSelection(projects),
-  );
+  const wizardOptionsWithPreSelectedProject = {
+    ...options,
+    preSelectedProject: {
+      project: selectedProject,
+      authToken,
+      selfHosted,
+    },
+  };
 
   const selectedTool = await traceStep('select-tool', askForUsedBundlerTool);
 
   Sentry.setTag('selected-tool', selectedTool);
 
   await traceStep('tool-setup', () =>
-    startToolSetupFlow(selectedTool, {
-      selfHosted,
-      orgSlug: selectedProject.organization.slug,
-      projectSlug: selectedProject.slug,
-      url: sentryUrl,
-      authToken: apiKeys.token,
-    }),
+    startToolSetupFlow(
+      selectedTool,
+      {
+        orgSlug: selectedProject.organization.slug,
+        projectSlug: selectedProject.slug,
+        selfHosted,
+        url: sentryUrl,
+        authToken,
+      },
+      wizardOptionsWithPreSelectedProject,
+    ),
   );
 
-  await traceStep('ci-setup', () => setupCi(apiKeys.token));
+  await traceStep('ci-setup', () => configureCI(selectedTool, authToken));
 
-  traceStep('outro', printOutro);
+  traceStep('outro', () =>
+    printOutro(
+      sentryUrl,
+      selectedProject.organization.slug,
+      selectedProject.id,
+    ),
+  );
 }
 
 async function askForUsedBundlerTool(): Promise<SupportedTools> {
   const selectedTool: SupportedTools | symbol = await abortIfCancelled(
     clack.select({
-      message: 'Which bundler or build tool are you using?',
+      message: 'Which framework, bundler or build tool are you using?',
       options: [
+        {
+          label: 'Angular',
+          value: 'angular',
+          hint: 'Select this option if you are using Angular.',
+        },
+        {
+          label: 'Create React App',
+          value: 'create-react-app',
+          hint: 'Select this option if you set up your app with Create React App.',
+        },
+        {
+          label: 'Next.js',
+          value: 'nextjs',
+          hint: 'Select this option if you want to set up source maps in a NextJS project.',
+        },
         {
           label: 'Webpack',
           value: 'webpack',
@@ -113,16 +159,12 @@ async function askForUsedBundlerTool(): Promise<SupportedTools> {
           hint: 'Configure source maps when using tsc as build tool',
         },
         {
-          label: 'Create React App',
-          value: 'create-react-app',
-          hint: 'Select this option if you set up your app with Create React App.',
-        },
-        {
           label: 'None of the above',
           value: 'sentry-cli',
           hint: 'This will configure source maps upload for you using sentry-cli',
         },
       ],
+      initialValue: await detectUsedTool(),
     }),
   );
 
@@ -132,6 +174,7 @@ async function askForUsedBundlerTool(): Promise<SupportedTools> {
 async function startToolSetupFlow(
   selctedTool: SupportedTools,
   options: SourceMapUploadToolConfigurationOptions,
+  wizardOptions: WizardOptions,
 ): Promise<void> {
   switch (selctedTool) {
     case 'webpack':
@@ -152,13 +195,78 @@ async function startToolSetupFlow(
     case 'create-react-app':
       await configureSentryCLI(options, configureCRASourcemapGenerationFlow);
       break;
+    case 'angular':
+      await configureSentryCLI(
+        options,
+        configureAngularSourcemapGenerationFlow,
+      );
+      break;
+    case 'nextjs':
+      await configureNextJsSourceMapsUpload(options, wizardOptions);
+      break;
     default:
       await configureSentryCLI(options);
       break;
   }
 }
 
-async function setupCi(authToken: string) {
+async function configureCI(
+  selectedTool: SupportedTools,
+  authToken: string,
+): Promise<void> {
+  const isUsingCI = await abortIfCancelled(
+    clack.select({
+      message: `Are you using a CI/CD tool to build and deploy your application?`,
+      options: [
+        {
+          label: 'Yes',
+          hint: 'I use a tool like Github Actions, Gitlab, CircleCI, TravisCI, Jenkins, ...',
+          value: true,
+        },
+        {
+          label: 'No',
+          hint: 'I build and deploy my application manually',
+          value: false,
+        },
+      ],
+      initialValue: true,
+    }),
+  );
+
+  Sentry.setTag('using-ci', isUsingCI);
+
+  const isCliBasedFlowTool = [
+    'sentry-cli',
+    'tsc',
+    'angular',
+    'create-react-app',
+  ].includes(selectedTool);
+
+  // some non-cli-based flows also use the .sentryclirc file
+  const usesSentryCliRc = selectedTool === 'nextjs';
+
+  const authTokenFile =
+    isCliBasedFlowTool || usesSentryCliRc
+      ? SENTRY_CLI_RC_FILE
+      : SENTRY_DOT_ENV_FILE;
+
+  if (!isUsingCI) {
+    clack.log.info(
+      `No Problem! Just make sure that the Sentry auth token from ${chalk.cyan(
+        authTokenFile,
+      )} is available whenever you build and deploy your app.`,
+    );
+    return;
+  }
+
+  if (isCliBasedFlowTool) {
+    await traceStep('ci-npm-script-setup', setupNpmScriptInCI);
+  }
+
+  await traceStep('ci-auth-token-setup', () => setupAuthTokenInCI(authToken));
+}
+
+async function setupAuthTokenInCI(authToken: string) {
   clack.log.step(
     'Add the Sentry authentication token as an environment variable to your CI setup:',
   );
@@ -199,19 +307,31 @@ SENTRY_AUTH_TOKEN=${authToken}
   }
 }
 
-function printOutro() {
+function printOutro(url: string, orgSlug: string, projectId: string) {
+  const pacMan = detectPackageManager() || 'npm';
+  const buildCommand = `'${pacMan}${pacMan === 'npm' ? ' run' : ''} build'`;
+
+  const urlObject = new URL(url);
+  urlObject.host = `${orgSlug}.${urlObject.host}`;
+  urlObject.pathname = '/issues/';
+  urlObject.searchParams.set('project', projectId);
+
+  const issueStreamUrl = urlObject.toString();
+
   const arrow = isUnicodeSupported() ? '→' : '->';
 
   clack.outro(`${chalk.green("That's it - everything is set up!")}
 
-   ${chalk.cyan(`Validate your setup with the following Steps:
+   ${chalk.cyan(`Test and validate your setup locally with the following Steps:
 
    1. Build your application in ${chalk.bold('production mode')}.
+      ${chalk.gray(`${arrow} For example, run ${chalk.bold(buildCommand)}.`)}
       ${chalk.gray(
         `${arrow} You should see source map upload logs in your console.`,
       )}
    2. Run your application and throw a test error.
-      ${chalk.gray(`${arrow} The error should be visible in Sentry.`)}
+      ${chalk.gray(`${arrow} The error should appear in Sentry:`)}
+      ${chalk.gray(`${arrow} ${issueStreamUrl}`)}
    3. Open the error in Sentry and verify that it's source-mapped.
       ${chalk.gray(
         `${arrow} The stack trace should show your original source code.`,